requirements.txt
# === Core LLM + LangChain Stack ===
openai>=1.54.3,<1.120.0
tiktoken>=0.8.0,<1.0.0
langchain==0.3.6
langchain-core>=0.3.14,<0.4.0
langchain-community==0.3.5
langchain-openai==0.2.5
langsmith==0.1.144
jiter==0.5.0
httpx==0.27.0

# === CrewAI + Agentic Tools ===
crewai==0.76.0
crewai-tools==0.76.0
# embedchain==0.1.127
instructor==1.6.3

# === Data + API Utilities ===
pyairtable==3.2.0
python-dotenv==1.1.1
pandas==2.3.3
pydantic==2.12.2
pydantic-core==2.41.4

# === FastAPI Server + WhatsApp Integration ===
fastapi==0.115.0
uvicorn==0.30.1
twilio==9.0.1
requests==2.32.3
python-multipart==0.0.20
async-timeout>=5.0.0,<6.0.0
# === Optional but helpful ===
nest-asyncio>=1.6.0
